sync

+++ cudacomp_MVMextractModesLoop.c
@@ -1003,37 +1003,37 @@ int  __attribute__((hot)) CUDACOMP_MVMextractModesLoop(
                 tstoptm = gmtime(&tstop.tv_sec);
 
                 if(data.signal_INT == 1) {
-                    sprintf(msgstring, "Received SIGINT at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGINT at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
                 if(data.signal_TERM == 1) {
-                    sprintf(msgstring, "Received SIGTERM at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGTERM at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
                 if(data.signal_ABRT == 1) {
-                    sprintf(msgstring, "Received SIGABRT at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGABRT at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
                 if(data.signal_BUS == 1) {
-                    sprintf(msgstring, "Received SIGBUS at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGBUS at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
                 if(data.signal_SEGV == 1) {
-                    sprintf(msgstring, "Received SIGSEGV at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGSEGV at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
                 if(data.signal_HUP == 1) {
-                    sprintf(msgstring, "Received SIGHUP at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGHUP at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
                 if(data.signal_PIPE == 1) {
-                    sprintf(msgstring, "Received SIGPIPE at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                    sprintf(msgstring, "Received SIGPIPE at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                     strncpy(processinfo->statusmsg, msgstring, 200);
                 }
 
@@ -1059,7 +1059,7 @@ int  __attribute__((hot)) CUDACOMP_MVMextractModesLoop(
                 clock_gettime(CLOCK_REALTIME, &tstop);
                 tstoptm = gmtime(&tstop.tv_sec);
 
-                sprintf(msgstring, "Received loop exit CTRL at %02d:%02d:%02d.%03ld", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, 0.000001*tstop.tv_nsec);
+                sprintf(msgstring, "Received loop exit CTRL at %02d:%02d:%02d.%03d", tstoptm->tm_hour, tstoptm->tm_min, tstoptm->tm_sec, (int) 0.000001*tstop.tv_nsec);
                 strncpy(processinfo->statusmsg, msgstring, 200);
 
                 processinfo->loopstat = 3; // clean exit
